Correct binding paths for chart widget
