+++ app/client/src/entities/Widget/utils.ts
@@ -81,6 +81,30 @@ export const getAllPathsFromPropertyConfig = (
             );
           }
         }
+        if (controlConfig.children) {
+          // Property in array structure
+          const basePropertyPath = controlConfig.propertyName;
+          const widgetPropertyValue = get(widget, basePropertyPath);
+          widgetPropertyValue.forEach(
+            (arrayPropertyValue: any, index: number) => {
+              const arrayIndexPropertyPath = `${basePropertyPath}[${index}]`;
+              controlConfig.children.forEach((childPropertyConfig: any) => {
+                const childArrayPropertyPath = `${arrayIndexPropertyPath}.${childPropertyConfig.propertyName}`;
+                if (
+                  childPropertyConfig.isBindProperty &&
+                  !childPropertyConfig.isTriggerProperty
+                ) {
+                  bindingPaths[childArrayPropertyPath] = true;
+                } else if (
+                  childPropertyConfig.isBindProperty &&
+                  childPropertyConfig.isTriggerProperty
+                ) {
+                  triggerPaths[childArrayPropertyPath] = true;
+                }
+              });
+            },
+          );
+        }
       });
     }
   });
