+++ app/client/src/widgets/ChartWidget/propertyConfig.ts
@@ -1,0 +1,127 @@
+import { ChartWidgetProps } from "widgets/ChartWidget/index";
+
+export default [
+  {
+    sectionName: "General",
+    children: [
+      {
+        helpText: "Adds a title to the chart",
+        placeholderText: "Enter title",
+        propertyName: "chartName",
+        label: "Title",
+        controlType: "INPUT_TEXT",
+        isBindProperty: true,
+        isTriggerProperty: false,
+      },
+      {
+        helpText: "Changes the visualisation of the chart data",
+        propertyName: "chartType",
+        label: "Chart Type",
+        controlType: "DROP_DOWN",
+        options: [
+          {
+            label: "Line Chart",
+            value: "LINE_CHART",
+          },
+          {
+            label: "Bar Chart",
+            value: "BAR_CHART",
+          },
+          {
+            label: "Pie Chart",
+            value: "PIE_CHART",
+          },
+          {
+            label: "Column Chart",
+            value: "COLUMN_CHART",
+          },
+          {
+            label: "Area Chart",
+            value: "AREA_CHART",
+          },
+        ],
+        isJSConvertible: true,
+        isBindProperty: true,
+        isTriggerProperty: false,
+      },
+      {
+        helpText: "Populates the chart with the data",
+        propertyName: "chartData",
+        placeholderText: 'Enter [{ "x": "val", "y": "val" }]',
+        label: "Chart Data",
+        controlType: "CHART_DATA",
+        hidden: (props: ChartWidgetProps) =>
+          props && props.chartType === "PIE_CHART",
+        isBindProperty: false,
+        isTriggerProperty: false,
+        children: [
+          {
+            helpText: "Series Name",
+            propertyName: "seriesName",
+            label: "Series Name",
+            controlType: "INPUT_TEXT",
+            isBindProperty: true,
+            isTriggerProperty: false,
+          },
+          {
+            helpText: "Series data",
+            propertyName: "data",
+            label: "Series Data",
+            controlType: "INPUT_TEXT_AREA",
+            isBindProperty: true,
+            isTriggerProperty: false,
+          },
+        ],
+      },
+      {
+        helpText: "Specifies the label of the x-axis",
+        propertyName: "xAxisName",
+        placeholderText: "Enter label text",
+        label: "x-axis Label",
+        controlType: "INPUT_TEXT",
+        isBindProperty: true,
+        isTriggerProperty: false,
+      },
+      {
+        helpText: "Specifies the label of the y-axis",
+        propertyName: "yAxisName",
+        placeholderText: "Enter label text",
+        label: "y-axis Label",
+        controlType: "INPUT_TEXT",
+        isBindProperty: true,
+        isTriggerProperty: false,
+      },
+      {
+        helpText: "Enables scrolling inside the chart",
+        propertyName: "allowHorizontalScroll",
+        label: "Allow horizontal scroll",
+        controlType: "SWITCH",
+        isBindProperty: false,
+        isTriggerProperty: false,
+      },
+      {
+        propertyName: "isVisible",
+        label: "Visible",
+        helpText: "Controls the visibility of the widget",
+        controlType: "SWITCH",
+        isJSConvertible: true,
+        isBindProperty: true,
+        isTriggerProperty: false,
+      },
+    ],
+  },
+  {
+    sectionName: "Actions",
+    children: [
+      {
+        helpText: "Triggers an action when the chart data point is clicked",
+        propertyName: "onDataPointClick",
+        label: "onDataPointClick",
+        controlType: "ACTION_SELECTOR",
+        isJSConvertible: true,
+        isBindProperty: true,
+        isTriggerProperty: true,
+      },
+    ],
+  },
+];
